requirements.txt
# AI Stock Trading Bot - Production Dependencies
# Install with: pip install -r requirements.txt

# Core Trading APIs
alpaca-trade-api>=3.0.0
python-dotenv>=1.0.0

# Market Data
yfinance>=0.2.28
alpha-vantage>=2.3.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.8.5

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Utilities
pytz>=2023.3
schedule>=1.2.0

# Technical Analysis
pandas-ta>=0.3.14b0

# Google Trends
pytrends==4.9.2

# Telegram Integration
python-telegram-bot>=20.0

# Database (Optional)
sqlalchemy>=2.0.0

# Date/Time
python-dateutil>=2.8.2

# Alternative Data Aggregator
tabulate>=0.9.0  # Required for pandas DataFrame markdown export

# ============================================================================
# Phase 2 Enhancements (New Dependencies)
# ============================================================================

# Anthropic Claude API (for debate system)
anthropic>=0.18.0

# Configuration Management
pyyaml>=6.0.1

# Async Support
asyncio>=3.4.3  # For async operations in Phase 2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Options Analysis (scipy already included above)
# Note: yfinance already included for options chain data

# Insider Trading Data
sec-api>=1.0.17  # SEC EDGAR API for insider trades (optional)

# Financial Datasets API (optional but recommended for real-time options flow)
# No package needed - uses aiohttp for API calls

# Additional utilities for Phase 2
tqdm>=4.66.0  # Progress bars for batch operations
colorama>=0.4.6  # Colored terminal output
